refactor: Add sessionId to movie/TV show details API calls
This commit modifies the `getMovieDetailsById` and `getTvShowDetailsById`
functions in the remote data sources and their corresponding API service
interfaces to include an optional `sessionId` parameter.

This change is necessary to fetch user-specific data, such as ratings,
when retrieving details for movies and TV shows.

The `AuthenticationLocalDataSourceImpl` has also been updated to return
an empty string instead of throwing an `UnauthorizedException` when the
session ID is not found, allowing for unauthenticated details fetching.

Additionally, the `MovieRepositoryImpl` and `TvShowRepositoryImpl` have been
updated to cache movie and TV show categories respectively if they are not
already cached locally. This ensures that category information is available
when saving movies/TV shows with their associated categories.