feature: add loading state to game screens
This commit introduces a loading state to the following game screens:
- `GuessReleaseYearGameScreen.kt`
- `GuessByPosterScreen.kt`
- `GuessCharacterScreen.kt`
- `GuessGenreScreen.kt`

A `LoadingContainer` is displayed with a fade-in/fade-out animation while the game questions are being fetched. The main game content is shown with a similar animation once the questions are available.

Additionally, this commit removes some `Log.e` statements from `MovieRemoteDataSourceImpl.kt`.